reworking directory structure with martins changes and now have tests that will run (even though integration tests arent passing).  Some of it is hacky and should be redone with proper python package/modules/namespacing
