--- requirements.txt
@@ -1,7 +1,0 @@
-PyQt5==5.10
-sip==4.19.7
-requests>=2.22.0
-psutil>=5.4.7
-behave>=1.2.6
-fbs==0.8.3
-black>=19.3b0
