+++ src/integration_tests/steps/step_login.py
@@ -5,7 +5,8 @@ from PyQt5.QtCore import Qt
 
 from behave import use_step_matcher, when, then
 
-from test.steps.step_helpers import assert_equal, assert_is_not, assert_is
+from src.integration_tests.steps.step_helpers import assert_equal, assert_is_not
+
 use_step_matcher("re")
